+++ Raspberry Pi/README.md
@@ -6,17 +6,17 @@ This file should contain documentation for the Raspberry Pi code.
 _Add virtualenv instructions here_
 
 ## Class overview
-![class-diagram](./class_diagram_v1.png)
+![class-diagram](./img/class_diagram_v1.png)
 
-#### rpi\_main.py - ThreadsafeSerialWriter
+#### microcontrollerIO.py - ThreadsafeSerialWriter
 Writes to a serial port in a thread safe manner. If two threads try to write to the same port, then the messages will be send properly without getting jumbled together. There should ever only be one instance of this class per serial port. All objects that write to the same port should share the ThreadSafeSerialWriter instance. 
 
 The `write` method takes a subject and message, and sends it to a serial port
 
-#### rpi\_main.py - SerialReader
+#### microcontrollerIO.py - SerialReader
 Reads a data of the form `XXXMESSAGE;` from the serial port. The data will be split into the subject and message (not including the semicolon) and will be sent to the `State`.
 
-#### rpi\_main.py - State
+#### state.py - State
 Contains the state of the all the sensors. There should only ever be one instance of this.
         
 A state has handlers,which should be thought of as `reacting` to messages from the serial port.
@@ -24,8 +24,8 @@ A state has handlers,which should be thought of as `reacting` to messages from t
 For instance, the `_handle_compass` method is called every time a message with subject CP is recieved, and it reacts by updating the `compass_angle` variable. Handlers should not be hugely complex, and should never need to be explicitly called. More complicated logic (autopilot, manual control) should be handled by the Controller, which can read from the State.
 
 
-#### rpi\_main.py - Controller
-Should eventually contain the majority of the autopilot control logic. This class has access to the State, and ThreadSafeSerialWriter. This can be implemented in a seperate thread than the reader, and the State object will be shared across threads.
+#### automatic\_control.py - Controller
+This class has access to the ThreadSafeSerialWriter. It should provide abstracted methods that allow control of the boat (e.g. actuate winch)
 
 ## Bytes in Python
 Unlike languages like C or C++, a Python `str` is not treated as an array of 8-bit bytes. Python has two built-in types for handling bytes - `bytes` and `bytearray`. This means that if someone wants to send a string to any serial i/o, one must convert into `bytes`. There are multiple ways to do this but for consistency, one should use the `encode()` method.
