+++ Raspberry Pi/automatic_control.py
@@ -1,0 +1,46 @@
+import logging
+import time
+
+
+class Controller:
+    """
+    Class used for direct control of the sailboat.
+
+    Args:
+        writer (ThreadsafeSerialWriter): The writer to send the messages too
+    """
+    def __init__(self, writer):
+        self.writer = writer
+
+    def actuate_winch(self, value):
+        assert 0 <= value <= 100
+        logging.info("Actuating winch: {}".format(value))
+        self.writer.write(b'00', str(value).encode('utf-8'))
+
+    def actuate_rudder(self, value):
+        assert 0 <= value <= 180
+        logging.info("Actuating rudder: {}".format(value))
+        self.writer.write(b'00', str(value).encode('utf-8'))
+
+
+def startAutomaticControl(state):
+    """
+    Code for controlling the boat based on the state object. The state object
+    contains up to date information of all the sensor readings.
+
+    It's highly important to know that this function does NOT try to modify the
+    state object, as there aren't any threadsafe guarantees. It should only
+    read from the state, never modifying its attributes.
+    """
+
+    logging.info("Starting automatic control thread.")
+    controller = Controller(state.writer)
+
+    while True:
+        if state.rpi_autopilot_enabled:
+            # Do something. E.g.
+            #
+            # controller.actuate_winch(22)
+            pass
+
+        time.sleep(0.5)
